requirements.txt
asgiref==3.6.0
dj-rest-auth==4.0.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
